pyproject.toml
[tool.poetry]
name = "python-substack"
version = "0.1.8"
description = "A Python wrapper around the Substack API."
authors = ["Paolo Mazza <mazzapaolo2019@gmail.com>"]
license = "MIT"
packages = [
    { include = "substack" }
]

readme = "README.md"

repository = "https://github.com/ma2za/python-substack"
homepage = "https://github.com/ma2za/python-substack"

keywords = ["substack"]

[tool.poetry.dependencies]
python = "^3.7"

requests = "^2.28.1"
python-dotenv = "^0.21.0"
PyYAML = "^6.0"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"